Add Navigate, useNavigate, and nest Route on Post
To demo how we can nest Routes, add a button to render on Post
and onClick useNavigate to reroute to '/about', and then to demo
if a res.status === 404, use <Navigate to='/notfound' /> to redirect
to a '404 Not Found' page.

For nesting Routes within a routed component:

- App.js
Add a '*' to '/post' route path <Route path='/post' />
like ('/post/*'), so anything after will still Route and
render <Post /> comp.

- Post.jsx
Add the <Route path={'/show'} /> which will simply nest after
the 'post' route like '/post/show'